npf/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
django-tinymce==4.0.0
django-unfold==0.22.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.11.0
